+++ src/Drivers/MDS/NlpMdsEx1.hpp
@@ -450,8 +450,12 @@ public:
                           bool& duals_avail,
                           double* z_bndL0,
                           double* z_bndU0,
-                          double* lambda0)
+                          double* lambda0,
+                          bool& slacks_avail,
+                          double* ineq_slack)
   {
+    slacks_avail = false;
+
     if(sol_x_ && sol_zl_ && sol_zu_ && sol_lambda_) {
 
       duals_avail = true;