+++ src/Drivers/PriDec/NlpPriDecEx1.cpp
@@ -149,9 +149,12 @@ bool PriDecEx1::get_starting_point(const size_type& n,
                                    bool& duals_avail,
                                    double* z_bndL0, 
                                    double* z_bndU0,
-                                   double* lambda0)
+                                   double* lambda0,
+                                   bool& slacks_avail,
+                                   double* ineq_slack)
 {
   duals_avail = false;
+  slacks_avail = false;
   return false;
 };