+++ src/LinAlg/hiopLinAlgFactory.cpp
@@ -69,8 +69,8 @@
 #include <hiopVectorRaja.hpp>
 #include <hiopMatrixRajaDense.hpp>
 #include <hiopMatrixRajaSparseTriplet.hpp>
-#include <hiopMatrixSparseCsrCuda.hpp>
 #endif // HIOP_USE_RAJA
+#include <hiopMatrixSparseCsrCuda.hpp>
 
 #include <hiopVectorIntSeq.hpp>
 #include <hiopVectorPar.hpp>