+++ src/LinAlg/hiopVectorRajaPar.cpp
@@ -1988,9 +1988,14 @@ bool hiopVectorRajaPar::isfinite_local() const
  * 
  * @pre Vector data was moved from the memory space to the host mirror.
  */
-void hiopVectorRajaPar::print(FILE* file, const char* msg/*=NULL*/, int max_elems/*=-1*/, int rank/*=-1*/) const
+void hiopVectorRajaPar::print(FILE* file/*=nullptr*/, const char* msg/*=nullptr*/, int max_elems/*=-1*/, int rank/*=-1*/) const
 {
-  int myrank=0, numranks=1; 
+  int myrank=0, numranks=1;
+
+  if(nullptr == file) {
+    file = stdout;
+  }
+
 #ifdef HIOP_USE_MPI
   if(rank >= 0) {
     int err = MPI_Comm_rank(comm_, &myrank); assert(err==MPI_SUCCESS);
@@ -2002,7 +2007,7 @@ void hiopVectorRajaPar::print(FILE* file, const char* msg/*=NULL*/, int max_elem
     if(max_elems>n_local_)
       max_elems=n_local_;
 
-    if(NULL==msg)
+    if(nullptr==msg)
     {
       std::stringstream ss;
       ss << "vector of size " << n_ << ", printing " << max_elems << " elems ";
@@ -2021,16 +2026,21 @@ void hiopVectorRajaPar::print(FILE* file, const char* msg/*=NULL*/, int max_elem
     }    
     fprintf(file, "=[");
     max_elems = max_elems >= 0 ? max_elems : n_local_;
-    for(int it=0; it<max_elems; it++)
+    for(int it=0; it<max_elems; it++) {
       fprintf(file, "%22.16e ; ", data_host_[it]);
+    }
     fprintf(file, "];\n");
   }
 }
 
+
 void hiopVectorRajaPar::print() const
 {
-  assert(nullptr != const_cast<hiopVectorRajaPar*>(this));
-  (const_cast<hiopVectorRajaPar*>(this))->copyFromDev();
+  //remove away const to be able to get the vec from the device
+  hiopVectorRajaPar* thisc = const_cast<hiopVectorRajaPar*>(this);
+  assert(thisc);
+
+  thisc->copyFromDev();
   for(index_type it=0; it<n_local_; it++) {
     printf("vec [%d] = %1.16e\n",it+1,data_host_[it]);
   }